changed type of for-loop variable from (size_t) to (singed long long), since unsigned numbers never get negative!
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@2026 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Base/TextBase.cpp
@@ -393,7 +393,7 @@ void TextBase :: makeLineDist()
     {
       // now in the other direction:
       // [-n ... -2 -1 0]
-	  size_t j;
+	  signed long long j;
       for(j=m_theText.size()-1; j>=0; j--){
         m_lineDist.push_back(-j);
       }
